Crate mina_crypto

Contains definitions of common simple crypto primitives used in the Mina protocol This currently includes keys and hashes along with their encodings

Modules

base58

base64

hash

hex

prelude

Macros

impl_bs58

impl_bs58_for_binprot

impl_bs58_full

impl_from_for_hash