Module mina_crypto::base58

Modules

decode

Functions for decoding Base58 encoded strings.

encode

Functions for encoding into Base58 encoded strings.

version_bytes

Macros

impl_bs58

impl_bs58_for_binprot

impl_bs58_full

Enums

Error

Traits

Base58Encodable

Functions

decode

Setup decoder for the given string using the [default alphabet][Alphabet::DEFAULT].

encode

Setup encoder for the given bytes using the [default alphabet][Alphabet::DEFAULT].